requirements.txt

torch>=1.10.0
torchvision>=0.11.0
transformers>=4.21.0
fastapi>=0.78.0
uvicorn>=0.17.0
pillow>=8.4.0
pyyaml>=6.0
